// server/interface/request.interface.ts
export interface AuthUrlParams {
  reURL: string
  channelCode: BankCode
}

export interface AccountRechargePageListParams extends ListPage {
  channelCode: BankCode
}

export interface SaveAccountRechargeParams {
  rechargeMoney: number
  reURL: string
  channelCode: BankCode
}
export interface UserLoginParams {
  loginRedirectStrategy?: string
  channelCode: BankCode
}

export interface AccountCashPageListParams extends ListPage {
  channelCode: BankCode
}
export interface QueryCashAccountParams {
  channelCode: BankCode
}

export interface SaveAccountCashParams {
  cashTotal: number
  isIntegrals: number
  reURL: string
  channelCode: BankCode
}
// ...
export interface SaveTenderBorrowParams {
  tenderType: number
  borrowNo: number
  tenderAmount: number
  ddjId?: number
  jxjId?: number
  reURL: string
  channelCode: BankCode
}